Clearing existing text from textView so that it's blank when the app is launched.
Add "s" from 2nd tapped so that it prints out
by adding if statement.
"The button got tapped 1 time"
"The button got tapped 2 times"
"The button got tapped 3 times" so on.

Make TextView scrollable manually.

And finally change the code so that a blank line is not printed before the messages, but is printed after them.